Refactor classes to follow current pattern
