+++ ElvargServer/src/main/java/com/elvarg/game/system/npc/impl/EmblemTraderInteraction.java
@@ -1,11 +1,11 @@
-package com.elvarg.game.content.combat.bountyhunter;
+package com.elvarg.game.system.npc.impl;
 
 import com.elvarg.game.entity.impl.npc.NPC;
 import com.elvarg.game.entity.impl.player.Player;
 import com.elvarg.game.model.container.shop.ShopManager;
 import com.elvarg.game.model.dialogues.builders.impl.EmblemTraderDialogue;
 import com.elvarg.game.system.InteractIds;
-import com.elvarg.game.system.NPCInteraction;
+import com.elvarg.game.system.npc.NPCInteraction;
 import com.elvarg.util.ShopIdentifiers;
 
 import static com.elvarg.util.NpcIdentifiers.*;
